veist_py/requirements.txt
discord.py
python-dotenv
pytest